# specs/001-supervised-autonomous-loop/contracts/http-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: Moonshift Supervised Autonomous Loop API
  version: 1.0.0-planned
  description: >-
    Loopback-only Iteration 0 contract for the single supervisor. This is a planning artifact,
    not a shipped or supported API.
servers:
  - url: http://127.0.0.1:{port}
    description: Local evaluation control plane
    variables:
      port:
        default: "4310"
security:
  - LocalSupervisorSession: []
...
paths:
...
  /v1/session/bootstrap:
    post:
      tags: [Session]
      operationId: bootstrapSupervisorSession
      security: []
      summary: Exchange one short-lived, single-use owner-local bootstrap secret for a browser session.
      description: >-
        Available only on a loopback bind. The secret arrives from a browser URL fragment, is never
        accepted in a query string, and is invalidated whether the exchange succeeds or fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bootstrapSecret]
              properties:
                bootstrapSecret:
                  type: string
                  minLength: 43
                  maxLength: 256
      responses:
        "204":
          description: Session established; response body is empty.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SupervisorSessionCookie"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/projects:
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create one durable fixture project from a supervisor objective.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectView"
        "200":
          description: The idempotency key already created this project; the original result is returned.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectView"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/projects/{projectId}:
    get:
      tags: [Projects]
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Current durable project projection.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectView"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects/{projectId}/events:
    get:
      tags: [Projects]
      operationId: streamProjectEvents
      summary: Stream an at-least-once projection ordered by project sequence.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: Last-Event-ID
          in: header
          required: false
          description: Last completely processed project sequence as a decimal string.
          schema:
            type: string
            pattern: "^[0-9]+$"
      responses:
        "200":
          description: >-
            Server-sent event stream. Each event id is the project sequence and each data field is a
            Project Event Envelope from event-envelope.schema.json.
          headers:
            Cache-Control:
              schema:
                const: no-cache
          content:
            text/event-stream:
              schema:
                type: string
        "409":
          description: >-
            Cursor is older than retained event history; reload ProjectView, including its complete
            durable-source presence projection, before opening a new event stream.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Problem"
                  - type: object
                    properties:
                      code:
                        const: EVENT_CURSOR_EXPIRED
...
components:
  securitySchemes:
    LocalSupervisorSession:
      type: apiKey
      in: cookie
      name: moonshift_supervisor_session
      description: >-
        Opaque host-only HttpOnly SameSite=Strict loopback evaluation session. It is never a provider
        credential and must not be exposed to JavaScript or logs.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable semantic command identity; 16 to 128 visible ASCII characters.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[!-~]+$"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    ETag:
      description: Strong aggregate-version ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    SupervisorSessionCookie:
      description: >-
        Host-only loopback session cookie with HttpOnly, SameSite=Strict, and Path=/ attributes; its
        opaque value must never be logged. Secure is additionally required when the loopback profile
        is served over HTTPS.
      schema:
        type: string
...
  responses:
...
    Problem:
      description: Safe problem detail with no stack, credential, or private reasoning.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [objective, fixtureScenario]
      properties:
        objective:
          type: string
          minLength: 1
          maxLength: 4000
        fixtureScenario:
          type: string
          enum:
            - PASS
            - EVIDENCE_FAIL
            - APPROVAL_REJECT
            - INTERRUPT_BEFORE_EFFECT
            - INTERRUPT_DURING_EFFECT
            - INTERRUPT_AFTER_EFFECT
...
    ProjectView:
      type: object
      additionalProperties: false
      required: [projectId, objective, status, version, personas, specialists, presences, channels, tasks, dependencies, capacity, lastSequence]
      properties:
        projectId:
          type: string
          format: uuid
        objective:
          type: string
        status:
          $ref: "#/components/schemas/ProjectState"
        version:
          type: integer
          minimum: 1
        personas:
          type: array
          description: >-
            Complete persistent persona identities; temporary specialists are projected separately.
            Policy limits ACTIVE personas to six while archived identities remain visible.
          minItems: 3
          items:
            $ref: "#/components/schemas/AgentSummary"
        specialists:
          type: array
          description: >-
            Current and archived task-scoped specialist identities needed to rebuild the organization
            view. Policy limits ACTIVE specialists to eight while archived identities remain visible.
          items:
            $ref: "#/components/schemas/AgentSummary"
        presences:
          type: array
          description: >-
            Complete current projection for every active persona and specialist, rebuilt from durable
            records and leases rather than inferred from the event stream or a socket.
          minItems: 3
          maxItems: 14
          items:
            $ref: "#/components/schemas/PresenceView"
        channels:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/ChannelSummary"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/TaskDependencySummary"
        capacity:
          $ref: "#/components/schemas/CapacityView"
        lastSequence:
          type: integer
          minimum: 0
    ProjectState:
      type: string
      enum:
        - CREATING
        - ACTIVE
        - PAUSING
        - PAUSED
        - RESUMING
        - STOPPING
        - STOPPED
        - CANCELLING
        - COMPLETED
        - BLOCKED
        - FAILED
        - CANCELLED
    AgentSummary:
      type: object
      additionalProperties: false
      required: [agentId, kind, role, status, lineageId]
      properties:
        agentId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [PERSONA, SPECIALIST]
        role:
          type: string
        status:
          $ref: "#/components/schemas/AgentIdentityState"
        lineageId:
          type: string
          format: uuid
    AgentIdentityState:
      type: string
      description: Durable identity lifecycle state; runtime presence is represented separately.
      enum: [CREATED, ACTIVE, PAUSED, DISABLED, COMPLETED, FAILED, ARCHIVED]
    PresenceState:
      type: string
      enum:
        - IDLE
        - QUEUED
        - THINKING_PROVIDER_CALL
        - USING_TOOLS
        - WAITING_FOR_RUNNER
        - WAITING_FOR_AGENT
        - WAITING_FOR_APPROVAL
        - VERIFYING
        - BLOCKED
        - COMPLETED
        - FAILED
    PresenceView:
      type: object
      additionalProperties: false
      required: [agentId, state, sourceType, sourceId, updatedAt, activity]
      properties:
        agentId:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/PresenceState"
        sourceType:
          type: string
          enum: [IDENTITY, TASK, EXECUTION, TOOL_INVOCATION, APPROVAL, CAPACITY, VERIFICATION_EVALUATION]
        sourceId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        activity:
          type: string
          minLength: 1
          maxLength: 240
          description: Bounded observable activity only; never a prompt, transcript, or private reasoning.
    ChannelSummary:
      type: object
      additionalProperties: false
      required: [channelId, name, kind, depth, status]
      properties:
        channelId:
          type: string
          format: uuid
        parentChannelId:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
        name:
          type: string
        kind:
          type: string
          enum: [CATEGORY, CHANNEL, SUBCHANNEL]
        depth:
          type: integer
          minimum: 0
          maximum: 4
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
    TaskSummary:
      type: object
      additionalProperties: false
      required: [taskId, title, state, expectedRevision]
      properties:
        taskId:
          type: string
          format: uuid
        title:
          type: string
        state:
          $ref: "#/components/schemas/TaskState"
        assigneeAgentId:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
        expectedRevision:
          type: string
          pattern: "^[a-f0-9]{40,64}$"
    TaskDependencySummary:
      type: object
      additionalProperties: false
      required: [taskDependencyId, predecessorTaskId, successorTaskId, kind]
      properties:
        taskDependencyId:
          type: string
          format: uuid
        predecessorTaskId:
          type: string
          format: uuid
        successorTaskId:
          type: string
          format: uuid
        kind:
          const: BLOCKS
    TaskState:
      type: string
      enum:
        - PROPOSED
        - READY
        - QUEUED
        - RUNNING
        - WAITING_FOR_AGENT
        - WAITING_FOR_CAPACITY
        - WAITING_FOR_APPROVAL
        - BLOCKED
        - CLAIMED_COMPLETE
        - VERIFYING
        - VERIFIED
        - FAILED
        - CANCELLED
...
    CapacityView:
      type: object
      additionalProperties: false
      required: [activeCognitiveRuns, cognitiveRunLimit, activeRunnerJobs, runnerJobLimit]
      properties:
        activeCognitiveRuns:
          type: integer
          minimum: 0
        cognitiveRunLimit:
          type: integer
          minimum: 1
          maximum: 5
        activeRunnerJobs:
          type: integer
          minimum: 0
        runnerJobLimit:
          type: integer
          const: 1
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        detail:
          type: string
          maxLength: 2000
        correlationId:
          type: string
          format: uuid